requirements.txt
lightning
matplotlib>3.1, <3.9.0
omegaconf >=2.2.3, <2.4.0
hydra-core >=1.2.0, <1.4.0
jsonargparse[signatures] >=4.27.7, <=4.35.0
rich >=12.3.0, <13.6.0
torchmetrics[image]
wandb
transformers==4.44.2
diffusers==0.32.2
timm
opencv-python-headless
einops
open-clip-torch==2.12.0
kornia
av==12.3.0
sentence_transformers
lancedb==0.14.0
pandas
accelerate
imageio-ffmpeg
imageio
numpy==1.26.4